Functionality for module 11 done

+++ module11/assignment.cl
@@ -12,17 +12,19 @@
 //
 //    This is a simple kernel performing convolution.
 
+// FLS - Use floats instead of integers
+
 __kernel void convolve(
-	const __global  uint * const input,
-    __constant uint * const mask,
-    __global  uint * const output,
+	const __global float * const input,
+    __constant float * const mask,
+    __global  float * const output,
     const int inputWidth,
     const int maskWidth)
 {
     const int x = get_global_id(0);
     const int y = get_global_id(1);
 
-    uint sum = 0;
+    float sum = 0;
     for (int r = 0; r < maskWidth; r++)
     {
         const int idxIntmp = (y + r) * inputWidth + x;
